requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
amzqr==0.0.1
    # via -r requirements.in
bcrypt==3.2.0
    # via paramiko
bidict==0.21.4
    # via python-socketio
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   flask
    #   pip-tools
cryptography==36.0.2
    # via paramiko
flask==2.0.3
    # via
    #   -r requirements.in
    #   flask-socketio
    #   flask-sqlalchemy
flask-socketio==5.1.1
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
gevent==21.12.0
    # via
    #   -r requirements.in
    #   gevent-websocket
gevent-websocket==0.10.1
    # via -r requirements.in
greenlet==1.1.2
    # via
    #   gevent
    #   sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
imageio==2.16.1
    # via amzqr
itsdangerous==2.1.1
    # via flask
jinja2==3.0.3
    # via flask
markupsafe==2.1.1
    # via jinja2
marshmallow==3.15.0
    # via
    #   -r requirements.in
    #   webargs
numpy==1.22.3
    # via
    #   -r requirements.in
    #   amzqr
    #   imageio
    #   opencv-python
opencv-python==4.5.5.64
    # via -r requirements.in
packaging==21.3
    # via
    #   marshmallow
    #   webargs
paramiko==2.10.2
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pillow==9.0.1
    # via
    #   amzqr
    #   imageio
pip-tools==6.5.1
    # via -r requirements.in
pycparser==2.21
    # via cffi
pycryptodome==3.14.1
    # via -r requirements.in
pyjwt==2.3.0
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.7
    # via packaging
python-dotenv==0.19.2
    # via -r requirements.in
python-engineio==4.3.1
    # via python-socketio
python-socketio==5.5.2
    # via flask-socketio
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via
    #   bcrypt
    #   paramiko
smart-getenv==1.1.0
    # via -r requirements.in
sqlalchemy==1.4.32
    # via flask-sqlalchemy
tomli==2.0.1
    # via pep517
urllib3==1.26.8
    # via requests
webargs==8.1.0
    # via -r requirements.in
werkzeug==2.0.3
    # via flask
wheel==0.37.1
    # via pip-tools
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
